Break out modules into files parseable by Jest
