+++ src/classes.js
@@ -1,0 +1,84 @@
+/* DEFINE CLASSES */
+
+function Tile(selector, currentValue=null, wasJustMerged=false, wasJustSpawned=false, previousValueMvLen=null) {
+  this.currentValue = currentValue;
+  this.wasJustMerged = wasJustMerged;
+  this.wasJustSpawned = wasJustSpawned;
+  this.previousValueMvLen = previousValueMvLen;
+  this.selector = selector;
+}
+
+function Board() {
+  this.hasChanged = false;
+  this.matrix = [];
+  for (let row = 0; row < 4; row++) {
+    this.matrix[row] = [];
+    for (let column = 0; column < 4; column++) {
+      this.matrix[row].push(new Tile(`#r${row}c${column}`));
+    }
+  }
+  this.spawnTiles = (howMany, isItTheOneAlready = false) => {
+    for (let i = 0; i < howMany; i++) {
+      let emptyTiles = [];
+      for (let row of this.matrix) {
+        emptyTiles.push(...row.filter(tile => !tile.currentValue))
+      }
+      let choiceIndex = Math.floor(Math.random() * emptyTiles.length);
+      emptyTiles[choiceIndex].currentValue =
+          isItTheOneAlready
+              ? 1
+              : Math.random() < 0.9
+              ? 2
+              : 4;
+      emptyTiles[choiceIndex].wasJustSpawned = true;
+    }
+  };
+  this.resetAnimationProperties = () => {
+    for (let row of this.matrix) {
+      for (let tile of row) {
+        tile.wasJustMerged = false;
+        tile.wasJustSpawned = false;
+        tile.previousValueMvLen = null;
+      }
+    }
+  };
+  this.hasChanged = () => {
+    for (let row of this.matrix) {
+      for (let tile of row) {
+        if (tile.previousValueMvLen || tile.wasJustMerged) {
+          return true;
+        }
+      }
+    }
+    return false;
+  };
+  this.gameStatus = () => {
+    let hasEmptySpots;
+    for (let row of this.matrix) {
+      for (let tile of row) {
+        if (tile.currentValue === 2048) {
+          return 'won';
+        }
+        if (tile.currentValue === null) {
+          hasEmptySpots = true;
+        }
+      }
+    }
+    if (hasEmptySpots) {
+      return 'ongoing'
+    }
+    for (let direction of ["up", "right", "down", "left"]) {
+      let testBoardCopy = createNextBoard(this, direction);
+      if (testBoardCopy.hasChanged()) {
+        return "ongoing"
+      }
+    }
+    return "lost";
+  }
+}
+
+
+module.exports = {
+  Tile,
+  Board,
+};
