+++ src/domManipulation.js
@@ -1,0 +1,33 @@
+/* DEFINE VIEW HANDLING FUNCTIONS */
+
+const updateMvAttributesInDOM = (board, direction) => {
+  for (let row of board.matrix) {
+    for (let tile of row) {
+      let tileElement = document.querySelector(tile.selector);
+      tileElement.setAttribute("data-mv-dir", direction);
+      tileElement.setAttribute("data-mv-len", tile.previousValueMvLen ? tile.previousValueMvLen : "");
+    }
+  }
+};
+
+const squashBoardInDOM = (nextBoard) => {
+  for (let row of nextBoard.matrix) {
+    for (let tile of row) {
+      let tileElement = document.querySelector(tile.selector);
+      tileElement.setAttribute("value", tile.currentValue);
+      tileElement.textContent = tile.currentValue;
+    }
+  }
+};
+
+const changeBackgroundInDOM = (color) => {
+  let body = document.querySelector('body');
+  body.setAttribute('style', `background-color: ${color}`);
+};
+
+
+module.exports = {
+  updateMvAttributesInDOM,
+  squashBoardInDOM,
+  changeBackgroundInDOM,
+};
